pyproject.toml
[tool.poetry]
name = "backend-project"
version = "0.1.0"
description = ""
authors = ["tetianaklius <tetyanaklyus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.4"
django-dotenv = "^1.4.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"

celery = "^5.4.0"

django-celery-results = "^2.5.1"
django-celery-beat = "^2.6.0"

#channels-redis = "^4.2.0"
#daphne = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
